+++ CHANGELOG.md
@@ -5,6 +5,7 @@ GAMSPy 0.13.3
 -------------
 - General
   - Change default solvers to 'CONOPT', 'CONVERT', 'CPLEX', 'GUSS', 'IPOPT', 'IPOPTH', 'KESTREL', 'NLPEC', 'PATH', and 'SHOT'
+  - Fix the version of gamspy_base when "gamspy update" command is being executed.
 - Testing
   - Update model library tests according to the new default solvers.